Extract catalog checkout finalize runtime
