+++ templates/operator/src/api/catalog-checkout.ts
@@ -1,5 +1,5 @@
 /**
- * Storefront checkout endpoint + workflow wiring.
+ * Storefront checkout endpoint.
  *
  * `POST /v1/public/catalog/checkout/start` — invoked by the
  * storefront's BookingJourney after the customer accepts the
@@ -17,12 +17,6 @@
  *                       the real opportunity creation.)
  *   - `hold`         → no payment workflow; staff broker the booking.
  *
- * Subscribes to `payment.completed`: when the Netopia webhook (card)
- * or the admin "Mark payment received" action (bank-transfer) fires
- * the event, runs the `checkout-finalize` workflow which transitions
- * the booking to `confirmed` and triggers contract + invoice
- * generation.
- *
  * See `docs/architecture/storefront-checkout-flow.md` Phases 3–5.
  */
 
@@ -33,41 +27,28 @@ import {
   createBookingPiiService,
   transitionBooking,
 } from "@voyantjs/bookings"
-import { bookingActivityLog, bookings } from "@voyantjs/bookings/schema"
-import {
-  type CheckoutFinalizeDeps,
-  type CheckoutFinalizeInput,
-  OWNED_SOURCE_KIND,
-  runCheckoutFinalize,
-} from "@voyantjs/catalog/booking-engine"
+import { bookings } from "@voyantjs/bookings/schema"
+import { OWNED_SOURCE_KIND } from "@voyantjs/catalog/booking-engine"
 import type { EventBus } from "@voyantjs/core"
 import {
   bookingItemTaxLines,
   type CreateInvoiceFromBookingInput,
   computeBookingItemTaxLine,
   financeService,
-  issueInvoiceFromBooking,
   issueProformaFromBooking,
   resolveBookingSellTaxRate,
 } from "@voyantjs/finance"
 import { parseJsonBody } from "@voyantjs/hono"
-import type { HonoBundle } from "@voyantjs/hono/plugin"
 import {
   NETOPIA_RUNTIME_CONTAINER_KEY,
   netopiaService,
   type ResolvedNetopiaRuntimeOptions,
 } from "@voyantjs/plugin-netopia"
-import {
-  beginWorkflowRun,
-  type WorkflowRunnerRegistry,
-  type WorkflowRunRecorder,
-} from "@voyantjs/workflow-runs"
-import { and, eq, isNull } from "drizzle-orm"
+import { and, eq } from "drizzle-orm"
 import type { PostgresJsDatabase } from "drizzle-orm/postgres-js"
 import type { Context, Hono } from "hono"
 import { z } from "zod"
 
-import { withDbFromEnv } from "./lib/db"
 import {
   getOperatorPaymentInstructions,
   getOperatorProfile,
@@ -157,26 +138,7 @@ export class CatalogCheckoutStartError extends Error {
   }
 }
 
-interface PaymentCompletedPayload {
-  bookingId: string | null
-  paymentSessionId?: string
-  paymentIntent?: "card" | "bank_transfer" | "hold" | "ticket_on_credit"
-  amountCents?: number
-  currency?: string
-  provider?: string | null
-}
-
-interface ContractDocumentGeneratedPayload {
-  contractId: string
-  contractStatus: string
-  attachmentId: string
-  attachmentKind: string
-  attachmentName: string
-}
-
-const ACCEPTANCE_MARKER_PREFIX = "__contract_acceptance__:"
-
-interface StoredAcceptance {
+interface CheckoutAcceptanceMetadata {
   templateId: string
   templateSlug: string
   acceptedAt: string
@@ -1780,39 +1742,6 @@ async function resolveBankTransferInstructions(
 }
 
 /**
- * Read the acceptance fingerprint stashed by the storefront from the
- * contract's own metadata. The pre-create flow (see
- * `persistAcceptanceDraftContract`) writes it to `contract.metadata.acceptance`
- * at /checkout/start time; this is the canonical home — no relay
- * through `bookings.internal_notes`.
- *
- * Older contracts created via the legacy marker-on-internal-notes
- * pattern fall back to the per-booking marker reader so existing
- * bookings finish their signature flow without manual re-entry.
- */
-function readContractAcceptance(
-  contractMetadata: unknown,
-  internalNotesFallback: string | null,
-): StoredAcceptance | null {
-  if (contractMetadata && typeof contractMetadata === "object") {
-    const meta = contractMetadata as Record<string, unknown>
-    if (meta.acceptance && typeof meta.acceptance === "object") {
-      return meta.acceptance as StoredAcceptance
-    }
-  }
-  if (!internalNotesFallback) return null
-  for (const line of internalNotesFallback.split("\n")) {
-    if (!line.startsWith(ACCEPTANCE_MARKER_PREFIX)) continue
-    try {
-      return JSON.parse(line.slice(ACCEPTANCE_MARKER_PREFIX.length)) as StoredAcceptance
-    } catch {
-      // Bad marker — try next line.
-    }
-  }
-  return null
-}
-
-/**
  * Pre-create (or update) a draft contract carrying the acceptance
  * fingerprint in `metadata.acceptance`. Called from /checkout/start
  * when the customer accepts the contract preview, BEFORE payment
@@ -1849,7 +1778,7 @@ async function persistAcceptanceDraftContract(
     return
   }
 
-  const acceptanceMetadata: StoredAcceptance = {
+  const acceptanceMetadata: CheckoutAcceptanceMetadata = {
     templateId: acceptance.templateId,
     templateSlug: acceptance.templateSlug,
     acceptedAt: acceptance.acceptedAt,
@@ -1894,560 +1823,3 @@ async function persistAcceptanceDraftContract(
     },
   })
 }
-
-async function persistAcceptanceSignature(
-  db: PostgresJsDatabase,
-  contractId: string,
-  eventBus?: EventBus,
-): Promise<void> {
-  const { contractsService } = await import("@voyantjs/legal/contracts")
-  const { contracts: contractsTable } = await import("@voyantjs/legal/contracts")
-  const [contract] = await db
-    .select()
-    .from(contractsTable)
-    .where(eq(contractsTable.id, contractId))
-    .limit(1)
-  if (!contract?.bookingId) return
-
-  const [booking] = await db
-    .select()
-    .from(bookings)
-    .where(eq(bookings.id, contract.bookingId))
-    .limit(1)
-  if (!booking) return
-
-  // Prefer the contract's own metadata (the new pre-create path);
-  // fall back to the legacy internal_notes marker for old contracts.
-  const acceptance = readContractAcceptance(contract.metadata, booking.internalNotes)
-  if (!acceptance) return
-
-  // Already signed? Idempotency comes from listSignatures.
-  const existing = await contractsService.listSignatures(db, contractId)
-  if (existing.length > 0) return
-
-  if (contract.status === "issued") {
-    const sent = await contractsService.sendContract(db, contractId, { eventBus })
-    if (sent.status !== "sent") {
-      console.warn(
-        `[catalog-checkout] could not send contract before acceptance signature for ${contractId}: ${sent.status}`,
-      )
-      return
-    }
-  }
-
-  // Prefer the booking's contact name when available so the signer
-  // line on the contract reads as the human, not the booking ref.
-  const contactName = [booking.contactFirstName, booking.contactLastName]
-    .filter(Boolean)
-    .join(" ")
-    .trim()
-  const signerName =
-    contactName ||
-    `Storefront customer${booking.bookingNumber ? ` (${booking.bookingNumber})` : ""}`
-
-  const result = await contractsService.signContract(
-    db,
-    contractId,
-    {
-      signerName,
-      signerEmail: booking.contactEmail ?? null,
-      method: "electronic" as const,
-      ipAddress: acceptance.clientIp ? acceptance.clientIp.slice(0, 64) : null,
-      userAgent: acceptance.userAgent ? acceptance.userAgent.slice(0, 500) : null,
-      metadata: {
-        source: "storefront-checkout",
-        templateId: acceptance.templateId,
-        templateSlug: acceptance.templateSlug,
-        acceptedAt: acceptance.acceptedAt,
-        acceptedMarketing: acceptance.acceptedMarketing,
-        renderedHtmlLength: acceptance.renderedHtmlLength,
-      },
-    } as never,
-    { eventBus },
-  )
-
-  if (result.status !== "signed") {
-    console.warn(
-      `[catalog-checkout] could not record acceptance signature for ${contractId}: ${result.status}`,
-    )
-    return
-  }
-
-  // Legacy cleanup: strip any leftover marker from internal_notes for
-  // contracts that came through the old relay path. New contracts
-  // (created via persistAcceptanceDraftContract) never had a marker
-  // written, so this branch no-ops.
-  if (booking.internalNotes?.includes(ACCEPTANCE_MARKER_PREFIX)) {
-    const cleanedNotes = booking.internalNotes
-      .split("\n")
-      .filter((line) => !line.startsWith(ACCEPTANCE_MARKER_PREFIX))
-      .join("\n")
-      .replace(/\n{3,}/g, "\n\n")
-      .trim()
-    await db
-      .update(bookings)
-      .set({
-        internalNotes: cleanedNotes.length > 0 ? cleanedNotes : null,
-        updatedAt: new Date(),
-      })
-      .where(eq(bookings.id, booking.id))
-  }
-}
-
-/**
- * Optional callback that generates (or fetches existing) the
- * contract PDF for a booking. Wired into `createCatalogCheckoutBundle`
- * by `app.ts` and forwarded into `CheckoutFinalizeDeps.generateContractPdf`
- * so the explicit `generate_contract_pdf` workflow step can run.
- *
- * Idempotent — if a contract document already exists for this
- * booking (because the legal subscriber on `booking.confirmed` won
- * the race, or because a prior run completed this step), the
- * implementation should return the existing ids rather than
- * generating a new contract.
- */
-export type CatalogCheckoutContractPdfGenerator = (input: {
-  env: CloudflareBindings
-  db: PostgresJsDatabase
-  eventBus: EventBus
-  bookingId: string
-}) => Promise<{ contractId: string; attachmentId: string } | null>
-
-/**
- * Build the deps object the `checkout-finalize` workflow expects,
- * binding to the supplied db / eventBus / recorder. Lifted out of
- * the event subscriber so the rerun / resume runners share the same
- * implementation — keeping rerun semantics in lockstep with the
- * normal "fired by payment.completed" path.
- */
-function buildCheckoutFinalizeDeps(
-  env: CloudflareBindings,
-  db: PostgresJsDatabase,
-  eventBus: EventBus,
-  recorder: WorkflowRunRecorder,
-  generateContractPdf?: CatalogCheckoutContractPdfGenerator,
-): CheckoutFinalizeDeps {
-  return {
-    db,
-    eventBus,
-    recorder: {
-      startStep: (name) => {
-        void recorder.startStep(name)
-      },
-      completeStep: (name, output) => {
-        void recorder.completeStep(name, output ?? null)
-      },
-      failStep: (name, error) => {
-        void recorder.failStep(name, error)
-      },
-    },
-    confirmBooking: async (bookingId) => {
-      const result = await bookingsService.confirmBooking(db, bookingId, {}, undefined, {
-        eventBus,
-      })
-      if (result.status === "ok") return
-
-      if (result.status === "hold_expired") {
-        const recovered = await bookingsService.recoverExpiredPaidBooking(
-          db,
-          bookingId,
-          { note: "Recovered after late payment completion" },
-          undefined,
-          { eventBus },
-        )
-        if (recovered.status === "ok") return
-        throw new Error(`checkout-finalize: late payment recovery failed (${recovered.status})`)
-      }
-
-      throw new Error(`checkout-finalize: booking confirmation failed (${result.status})`)
-    },
-    issueInvoice: async ({ bookingId, convertedFromInvoiceId }) => {
-      const [booking] = await db.select().from(bookings).where(eq(bookings.id, bookingId)).limit(1)
-      if (!booking) return null
-
-      const { bookingItems } = await import("@voyantjs/bookings/schema")
-      const items = await db
-        .select()
-        .from(bookingItems)
-        .where(eq(bookingItems.bookingId, bookingId))
-
-      const today = new Date().toISOString().slice(0, 10)
-      const dueDate = new Date(Date.now() + 14 * 24 * 60 * 60 * 1000).toISOString().slice(0, 10)
-
-      const invoice = await issueInvoiceFromBooking(
-        db,
-        {
-          bookingId,
-          issueDate: today,
-          dueDate,
-          invoiceType: "invoice",
-          notes: convertedFromInvoiceId
-            ? `Converted from proforma ${convertedFromInvoiceId}`
-            : null,
-        },
-        {
-          booking: {
-            id: booking.id,
-            bookingNumber: booking.bookingNumber,
-            personId: booking.personId,
-            organizationId: booking.organizationId,
-            sellCurrency: booking.sellCurrency,
-            baseCurrency: booking.baseCurrency,
-            fxRateSetId: null,
-            sellAmountCents: booking.sellAmountCents,
-            baseSellAmountCents: booking.baseSellAmountCents,
-          },
-          items: items.map((item) => ({
-            id: item.id,
-            title: item.title,
-            quantity: item.quantity,
-            unitSellAmountCents: item.unitSellAmountCents,
-            totalSellAmountCents: item.totalSellAmountCents,
-          })),
-        },
-        { eventBus },
-      )
-
-      if (invoice && convertedFromInvoiceId) {
-        await db
-          .update((await import("@voyantjs/finance")).invoices)
-          .set({ convertedFromInvoiceId })
-          .where(eq((await import("@voyantjs/finance")).invoices.id, invoice.id))
-      }
-
-      return invoice ? { invoiceId: invoice.id } : null
-    },
-    findProformaForBooking: async (bookingId) => {
-      const { invoices } = await import("@voyantjs/finance")
-      const [proforma] = await db
-        .select({ id: invoices.id })
-        .from(invoices)
-        .where(eq(invoices.bookingId, bookingId))
-        .limit(1)
-      return proforma ? { invoiceId: proforma.id } : null
-    },
-    generateContractPdf: generateContractPdf
-      ? async ({ bookingId }) => generateContractPdf({ env, db, eventBus, bookingId })
-      : undefined,
-    linkPaymentToInvoice: async ({ bookingId, invoiceId, paymentSessionId }) => {
-      // Find paid sessions targeting this booking that aren't yet
-      // linked to an invoice. There's usually exactly one (the
-      // checkout session the customer just paid), but the loop covers
-      // edge cases like split deposits paid through multiple sessions.
-      const { paymentSessions } = await import("@voyantjs/finance/schema")
-      const { financeService } = await import("@voyantjs/finance")
-      const paidSessions = await db
-        .select()
-        .from(paymentSessions)
-        .where(
-          and(
-            eq(paymentSessions.bookingId, bookingId),
-            eq(paymentSessions.status, "paid"),
-            isNull(paymentSessions.invoiceId),
-          ),
-        )
-
-      let firstPaymentId: string | null = null
-      let sessionsLinked = 0
-
-      for (const session of paidSessions) {
-        // 1. Back-link the session to the invoice for audit / future
-        //    queries (`SELECT * FROM payment_sessions WHERE invoice_id = ?`).
-        await db
-          .update(paymentSessions)
-          .set({ invoiceId, updatedAt: new Date() })
-          .where(eq(paymentSessions.id, session.id))
-
-        // 2. Record a `payments` row tying real money received to the
-        //    invoice. `financeService.createPayment` runs in a
-        //    transaction that recalculates `paid_cents` /
-        //    `balance_due_cents` and flips the invoice to "paid"
-        //    when the totals match.
-        const payment = await financeService.createPayment(db, invoiceId, {
-          amountCents: session.amountCents,
-          currency: session.currency,
-          paymentMethod: session.paymentMethod ?? "credit_card",
-          paymentInstrumentId: session.paymentInstrumentId ?? null,
-          paymentAuthorizationId: session.paymentAuthorizationId ?? null,
-          paymentCaptureId: session.paymentCaptureId ?? null,
-          status: "completed",
-          referenceNumber:
-            session.providerPaymentId ??
-            session.externalReference ??
-            session.providerSessionId ??
-            session.id,
-          paymentDate: (session.completedAt ?? new Date()).toISOString().slice(0, 10),
-          notes:
-            `Checkout-finalize linkage from session ${session.id}` +
-            (paymentSessionId && session.id !== paymentSessionId
-              ? ` (workflow input session: ${paymentSessionId})`
-              : ""),
-        })
-
-        // 3. Stamp the session with the payments.id so the operator
-        //    UI can resolve session → payment in one hop.
-        if (payment?.id) {
-          await db
-            .update(paymentSessions)
-            .set({ paymentId: payment.id, updatedAt: new Date() })
-            .where(eq(paymentSessions.id, session.id))
-          if (!firstPaymentId) firstPaymentId = payment.id
-        }
-        sessionsLinked++
-      }
-
-      return { paymentId: firstPaymentId, sessionsLinked }
-    },
-  }
-}
-
-interface DispatchCheckoutFinalizeParams {
-  env: CloudflareBindings
-  db: PostgresJsDatabase
-  eventBus: EventBus
-  /** Workflow input — booking + payment metadata. */
-  input: CheckoutFinalizeInput
-  /** Run-row metadata: trigger string, correlationId, tags. */
-  trigger: string
-  correlationId: string | null
-  tags: ReadonlyArray<string>
-  /** When set, run is recorded as a child of this parent run. */
-  parentRunId?: string | null
-  /** User who triggered the run (for audit). */
-  triggeredByUserId?: string | null
-  /** When set, run is resumed at this step. */
-  resumeFromStep?: string
-  /** Pre-seeded step results from the parent (resume path). */
-  seedResults?: Record<string, unknown>
-  /** Optional contract-PDF generator forwarded into deps. */
-  generateContractPdf?: CatalogCheckoutContractPdfGenerator
-}
-
-/**
- * Begin a `checkout-finalize` run, execute the workflow, and finalize
- * the run row. Used by the `payment.completed` subscriber and the
- * rerun/resume runners. Returns the new run id.
- *
- * Recorder failures don't mask workflow failures — `recorder.fail` is
- * fire-and-forget and we still propagate the original error to the
- * caller (the runner) for HTTP error reporting.
- */
-async function dispatchCheckoutFinalize(
-  params: DispatchCheckoutFinalizeParams,
-): Promise<{ runId: string }> {
-  const recorder = await beginWorkflowRun(params.db, {
-    workflowName: "checkout-finalize",
-    trigger: params.trigger,
-    correlationId: params.correlationId ?? null,
-    tags: [...params.tags],
-    input: params.input as unknown as Record<string, unknown>,
-    parentRunId: params.parentRunId ?? null,
-    triggeredByUserId: params.triggeredByUserId ?? null,
-    resumeFromStep: params.resumeFromStep ?? null,
-  })
-
-  // Manual rerun/resume — write a `system_action` activity row on
-  // the booking so the customer-facing booking detail page surfaces
-  // an entry like "checkout-finalize was manually resumed from
-  // issue_invoice by Mihai". The recorded `workflow_runs` row
-  // already exists (one above), but operators land on the booking
-  // page first and shouldn't have to cross-reference the dashboard
-  // to know a workflow was retried.
-  if (params.parentRunId) {
-    try {
-      const action = params.resumeFromStep ? "resumed" : "rerun"
-      const description = params.resumeFromStep
-        ? `Workflow checkout-finalize ${action} from step "${params.resumeFromStep}"`
-        : `Workflow checkout-finalize ${action}`
-      await params.db.insert(bookingActivityLog).values({
-        bookingId: params.input.bookingId,
-        actorId: params.triggeredByUserId ?? null,
-        activityType: "system_action",
-        description,
-        metadata: {
-          kind: "workflow_rerun",
-          workflowName: "checkout-finalize",
-          parentRunId: params.parentRunId,
-          newRunId: recorder.runId,
-          resumeFromStep: params.resumeFromStep ?? null,
-        },
-      })
-    } catch (err) {
-      console.warn("[catalog-checkout] failed to write rerun activity log", err)
-    }
-  }
-
-  // Resume path: write skipped step rows for everything before the
-  // resume target so the UI shows the full step list with the source
-  // of each output (parent run output, marked "skipped").
-  if (params.resumeFromStep && params.seedResults) {
-    for (const [stepName, output] of Object.entries(params.seedResults)) {
-      // Skip the synthetic deps step — it has no observable meaning
-      // in the dashboard, and re-recording it as "skipped" would
-      // confuse users.
-      if (stepName === "__deps") continue
-      await recorder.recordSkippedStep(
-        stepName,
-        output && typeof output === "object" ? (output as Record<string, unknown>) : null,
-      )
-    }
-  }
-
-  const deps = buildCheckoutFinalizeDeps(
-    params.env,
-    params.db,
-    params.eventBus,
-    recorder,
-    params.generateContractPdf,
-  )
-  try {
-    await runCheckoutFinalize(params.input, deps, {
-      skipUntil: params.resumeFromStep,
-      seedResults: params.seedResults,
-    })
-    await recorder.complete()
-    return { runId: recorder.runId }
-  } catch (err) {
-    console.error("[catalog-checkout] checkout-finalize workflow failed", err)
-    await recorder.fail(err)
-    throw err
-  }
-}
-
-/**
- * Bundle factory that subscribes to `payment.completed` (running the
- * `checkout-finalize` workflow) and `contract.document.generated`
- * (promoting the storefront's acceptance marker into a real
- * `contract_signatures` row), and registers the `checkout-finalize`
- * runner with the supplied {@link WorkflowRunnerRegistry} so the
- * dashboard's "Rerun" / "Resume" buttons work.
- *
- * The runner is declared `idempotency: "unsafe"` because a fresh
- * rerun would attempt to issue another invoice number. The dashboard
- * requires a confirm dialog before sending `confirm: true`. The
- * Resume path is always safe — it skips already-completed steps.
- */
-export function createCatalogCheckoutBundle(opts: {
-  workflowRunnerRegistry?: WorkflowRunnerRegistry
-  /**
-   * Hook the explicit `generate_contract_pdf` step in the workflow
-   * to a real implementation. Operators that wire this also rely on
-   * `autoGenerateContractForBooking` being idempotent (it is) so the
-   * legal package's `booking.confirmed` subscriber and this step
-   * coexist without double-generating.
-   *
-   * Omit to skip the step (returns `null`) — useful for deploys that
-   * don't want explicit-step recording, falling back to the
-   * subscriber-only path.
-   */
-  generateContractPdf?: CatalogCheckoutContractPdfGenerator
-}): HonoBundle {
-  return {
-    name: "catalog-checkout",
-    bootstrap: ({ bindings, eventBus }) => {
-      const env = bindings as CloudflareBindings
-      eventBus.subscribe<ContractDocumentGeneratedPayload>(
-        "contract.document.generated",
-        async ({ data }) => {
-          try {
-            await withDbFromEnv(env, async (rawDb) => {
-              const db = rawDb as unknown as PostgresJsDatabase
-              await persistAcceptanceSignature(db, data.contractId, eventBus)
-            })
-          } catch (err) {
-            console.error("[catalog-checkout] persistAcceptanceSignature failed", err)
-          }
-        },
-      )
-      eventBus.subscribe<PaymentCompletedPayload>("payment.completed", async ({ data }) => {
-        if (!data.bookingId) return
-        const bookingId = data.bookingId
-        try {
-          await withDbFromEnv(env, async (rawDb) => {
-            const db = rawDb as unknown as PostgresJsDatabase
-            await dispatchCheckoutFinalize({
-              env,
-              db,
-              eventBus,
-              input: {
-                bookingId,
-                paymentSessionId: data.paymentSessionId,
-                paymentIntent: data.paymentIntent,
-              },
-              trigger: "payment.completed",
-              correlationId: data.paymentSessionId ?? null,
-              tags: [
-                `bookingId:${bookingId}`,
-                ...(data.paymentSessionId ? [`paymentSessionId:${data.paymentSessionId}`] : []),
-                ...(data.paymentIntent ? [`paymentIntent:${data.paymentIntent}`] : []),
-              ],
-              generateContractPdf: opts.generateContractPdf,
-            })
-          })
-        } catch {
-          // dispatchCheckoutFinalize already logged + recorded the
-          // failure; swallow here so the event-bus callback doesn't
-          // bubble to the dispatch caller.
-        }
-      })
-
-      if (opts.workflowRunnerRegistry) {
-        opts.workflowRunnerRegistry.register({
-          name: "checkout-finalize",
-          idempotency: "unsafe",
-          description:
-            "Confirms the booking and issues the final invoice. A fresh rerun issues another invoice number; use Resume to retry from a failed step.",
-          rerun: async (rawInput, ctx) => {
-            const input = rawInput as CheckoutFinalizeInput | null
-            if (!input?.bookingId) {
-              throw new Error("checkout-finalize rerun: recorded input has no bookingId")
-            }
-            return withDbFromEnv(env, async (rawDb) => {
-              const db = rawDb as unknown as PostgresJsDatabase
-              return dispatchCheckoutFinalize({
-                env,
-                db,
-                eventBus,
-                input,
-                trigger: "manual.rerun",
-                correlationId: ctx.correlationId,
-                tags: [...ctx.tags, "rerun:true"],
-                parentRunId: ctx.parentRunId,
-                triggeredByUserId: ctx.triggeredByUserId,
-                generateContractPdf: opts.generateContractPdf,
-              })
-            })
-          },
-          resume: async (rawInput, ctx) => {
-            const input = rawInput as CheckoutFinalizeInput | null
-            if (!input?.bookingId) {
-              throw new Error("checkout-finalize resume: recorded input has no bookingId")
-            }
-            return withDbFromEnv(env, async (rawDb) => {
-              const db = rawDb as unknown as PostgresJsDatabase
-              return dispatchCheckoutFinalize({
-                env,
-                db,
-                eventBus,
-                input,
-                trigger: "manual.resume",
-                correlationId: ctx.correlationId,
-                tags: [...ctx.tags, "resume:true"],
-                parentRunId: ctx.parentRunId,
-                triggeredByUserId: ctx.triggeredByUserId,
-                resumeFromStep: ctx.resumeFromStep,
-                seedResults: ctx.seedResults,
-                generateContractPdf: opts.generateContractPdf,
-              })
-            })
-          },
-        })
-      }
-    },
-  }
-}
-
-/** @deprecated Kept for callers that still import the static bundle —
- *  use {@link createCatalogCheckoutBundle} so the rerun/resume
- *  runner is registered with the dashboard. */
-export const catalogCheckoutBundle = createCatalogCheckoutBundle({})
